+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -17,7 +17,6 @@ using Libplanet.Crypto;
 using Libplanet.Store;
 using Libplanet.Store.Trie;
 using Libplanet.Tests.Action;
-using Libplanet.Tests.Blocks;
 using Libplanet.Tests.Common.Action;
 using Libplanet.Tests.Store;
 using Libplanet.Tx;
@@ -93,49 +92,6 @@ namespace Libplanet.Tests.Blockchain
         public void Dispose()
         {
             _fx.Dispose();
-        }
-
-        [SkippableFact]
-        public void PerceiveBlock()
-        {
-            var blockA = new SimpleBlockExcerpt()
-            {
-                ProtocolVersion = BlockMetadata.CurrentProtocolVersion,
-                Index = 604665,
-                Hash = BlockHash.FromString(
-                    "4f612467ed79cb854d1901f131ccfc8a40bba89651e1a9e1dcea1287dd70d8ee"),
-                TotalDifficulty = 21584091240753,
-            };
-
-            DateTimeOffset timeA = DateTimeOffset.FromUnixTimeSeconds(1609426800);
-            BlockPerception perceptionA = _blockChain.PerceiveBlock(blockA, timeA);
-            Assert.True(blockA.ExcerptEquals(perceptionA));
-            Assert.Equal(timeA, perceptionA.PerceivedTime);
-
-            perceptionA = _blockChain.PerceiveBlock(blockA);
-            Assert.True(blockA.ExcerptEquals(perceptionA));
-            Assert.Equal(timeA, perceptionA.PerceivedTime);
-
-            var blockB = new SimpleBlockExcerpt
-            {
-                ProtocolVersion = BlockMetadata.CurrentProtocolVersion,
-                Index = 604664,
-                Hash = BlockHash.FromString(
-                    "9a87556f3198d8bd48300d2a6a5957d661c760a7fb72ef4a4b8c01c155b77e99"),
-                TotalDifficulty = 21584061959429,
-            };
-
-            DateTimeOffset timeBMin = DateTimeOffset.FromUnixTimeMilliseconds(
-                DateTimeOffset.UtcNow.ToUnixTimeMilliseconds());
-            BlockPerception perceptionB = _blockChain.PerceiveBlock(blockB);
-            DateTimeOffset timeBMax = DateTimeOffset.UtcNow;
-            Assert.True(blockB.ExcerptEquals(perceptionB));
-            Assert.InRange(perceptionB.PerceivedTime, timeBMin, timeBMax);
-
-            DateTimeOffset timeB = perceptionB.PerceivedTime;
-            perceptionB = _blockChain.PerceiveBlock(blockB);
-            Assert.True(blockB.ExcerptEquals(perceptionB));
-            Assert.Equal(timeB, perceptionB.PerceivedTime);
         }
 
         [SkippableFact]